vigilwolf-core/backend/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# HTTP and Web
requests>=2.31.0
httpx>=0.24.0
beautifulsoup4>=4.12.0

# Screenshot capture (Playwright preferred, Selenium as fallback)
playwright>=1.40.0
selenium>=4.15.0

# Background Tasks
apscheduler>=3.10.0

# Date/Time Handling
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
hypothesis>=6.92.0

# Existing VigilWolf Dependencies
fuzzysearch
python-whois

# Utilities
python-dotenv>=1.0.0